+++ eslint.config.js
@@ -22,6 +22,7 @@ export default [
     rules: {
       '@typescript-eslint/no-unused-vars': 'warn',
       'prettier/prettier': 'error',
+      'arrow-parens': ['error', 'always'],
     },
   },
 ];